+++ src/tui.rs
@@ -53,6 +53,12 @@ struct UISettings {
     error_text: String,
     error_timer: Option<Instant>,
     device_name: String,
+    waveform_window: f64,
+    // Used to flash control elements when the button is pressed
+    left_arrow_timer: Option<Instant>,
+    right_arrow_timer: Option<Instant>,
+    plus_sign_timer: Option<Instant>,
+    minus_sign_timer: Option<Instant>,
 }
 
 impl Default for UISettings {
@@ -69,6 +75,11 @@ impl Default for UISettings {
             error_text: String::new(),
             error_timer: None,
             device_name: String::new(),
+            waveform_window: 15.,
+            left_arrow_timer: None,
+            right_arrow_timer: None,
+            plus_sign_timer: None,
+            minus_sign_timer: None,
         }
     }
 }
@@ -124,9 +135,11 @@ impl Theme {
         fill_fields!(self.waveform.
             borders => fg,
             controls => fg,
+            controls_highlight => hl,
             labels => fg,
             playhead => hl,
-            time => fg,
+            current_time => fg,
+            total_duration => fg,
             waveform => fg,
             background => bg,
             highlight => hl,
@@ -156,6 +169,7 @@ impl Theme {
 
         fill_fields!(self.explorer.
             background => bg,
+            borders => fg,
             dir_foreground => fg,
             item_foreground => fg,
             highlight_dir_foreground => hl,
@@ -195,9 +209,11 @@ struct WaveformTheme {
     waveform: Option<Color>,
     playhead: Option<Color>,
     /// Current playing time and total duration
-    time: Option<Color>,
+    current_time: Option<Color>,
+    total_duration: Option<Color>,
     /// Buttons like <-, +, -, ->
     controls: Option<Color>,
+    controls_highlight: Option<Color>,
     labels: Option<Color>,
     /// Background of the chart
     background: Option<Color>,
@@ -247,6 +263,7 @@ struct DevicesTheme {
 #[derive(Deserialize)]
 struct ExplorerTheme {
     background: Option<Color>,
+    borders: Option<Color>,
     item_foreground: Option<Color>,
     highlight_item_foreground: Option<Color>,
     dir_foreground: Option<Color>,
@@ -277,8 +294,10 @@ impl Default for WaveformTheme {
             borders: Some(Color::Yellow),
             waveform: Some(Color::Yellow),
             playhead: Some(Color::LightRed),
-            time: Some(Color::Yellow),
+            current_time: Some(Color::Yellow),
+            total_duration: Some(Color::Yellow),
             controls: Some(Color::Yellow),
+            controls_highlight: Some(Color::LightRed),
             labels: Some(Color::Yellow),
             background: Some(Color::Black),
             highlight: Some(Color::LightRed),
@@ -320,7 +339,7 @@ impl Default for DevicesTheme {
     fn default() -> Self {
         Self {
             background: Some(Color::Black),
-            foreground: Some(Color::White),
+            foreground: Some(Color::Yellow),
         }
     }
 }
@@ -329,6 +348,7 @@ impl Default for ExplorerTheme {
     fn default() -> Self {
         Self {
             background: Some(Color::Black),
+            borders: Some(Color::Yellow),
             item_foreground: Some(Color::Yellow),
             highlight_item_foreground: Some(Color::LightRed),
             dir_foreground: Some(Color::Yellow),
@@ -447,7 +467,9 @@ impl App {
 
     fn set_theme(&mut self, theme: Theme) {
         // define styles
-        let s = Style::default().bg(theme.explorer.background.unwrap());
+        let s = Style::default()
+            .bg(theme.explorer.background.unwrap())
+            .fg(theme.explorer.borders.unwrap());
         let is = s.fg(theme.explorer.item_foreground.unwrap());
         let ihl = s.fg(theme.explorer.highlight_item_foreground.unwrap());
         let ds = s.fg(theme.explorer.dir_foreground.unwrap()).bold();
@@ -512,7 +534,8 @@ impl App {
         let _ct = s.fg(self.ui_settings.theme.waveform.controls.unwrap());
         let hl = s.fg(self.ui_settings.theme.waveform.highlight.unwrap());
         let pl = s.fg(self.ui_settings.theme.waveform.playhead.unwrap());
-        let tm = s.fg(self.ui_settings.theme.waveform.time.unwrap());
+        let ct = s.fg(self.ui_settings.theme.waveform.current_time.unwrap());
+        let td = s.fg(self.ui_settings.theme.waveform.total_duration.unwrap());
         let wv = s.fg(self.ui_settings.theme.waveform.waveform.unwrap());
         // playhead is just a function that looks like a vertical line
         let samples_in_one_ms = self.audio_file.sample_rate() / 1000;
@@ -524,52 +547,11 @@ impl App {
             ),
         ];
         if self.waveform.at_end {
-            // if at last 15 sec of the audion the playhead should move from the middle to the end of the chart
-            let total_samples = self.audio_file.mid_samples().len();
-            let chart_duration_seconds = 15.0; // make a var not to hard code and be able to to add resizing waveform window if needed
-            let chart_middle_seconds = chart_duration_seconds / 2.0;
-
-            // calculate the absolute sample position where the playhead starts scrolling from the middle of the chart to the end
-            // this is when playback enters the last `chart_middle_seconds` (default is 7.5s) of the total audio duration.
-            let scroll_start_absolute_samples = total_samples.saturating_sub(
-                (chart_middle_seconds * self.audio_file.sample_rate() as f64) as usize,
-            );
-
-            // calculate playhead's position relative to the start of this scroll phase
-            // since `self.waveform.playhead` is the absolute current playback position.
-            let relative_samples_in_scroll_phase = self
-                .waveform
-                .playhead
-                .saturating_sub(scroll_start_absolute_samples);
-
-            // map this relative sample position to the chart's X-axis range for the playhead.
-            // the conversion from samples to chart units (milliseconds) uses the same 1/samles_in_one_ms scale
-            // as other playhead positions in this function.
-            let mut chart_x_position = (chart_middle_seconds * 1000.)
-                + (relative_samples_in_scroll_phase as f64 / samples_in_one_ms as f64);
-
-            // Ensure the playhead does not exceed the chart's upper bound.
-            chart_x_position = f64::min(chart_x_position, chart_duration_seconds * 1000.);
-
+            let chart_x_position = self.get_relative_playhead_pos(samples_in_one_ms);
             playhead_chart = [(chart_x_position, 1.), (chart_x_position + 0.01, -1.)];
         } else if !self.waveform.at_zero {
             // if not at zero then place the playhead right at the middle of a chart
-            playhead_chart = [
-                (
-                    f64::min(
-                        self.waveform.playhead as f64 / samples_in_one_ms as f64,
-                        1000. * 7.5,
-                    ),
-                    1.,
-                ),
-                (
-                    f64::min(
-                        self.waveform.playhead as f64 / samples_in_one_ms as f64,
-                        1000. * 7.5,
-                    ) + 0.01,
-                    -1.,
-                ),
-            ];
+            playhead_chart = self.get_middle_playhead_pos(samples_in_one_ms);
         }
         if self.settings.mode != Mode::Player {
             playhead_chart = [(-1., -1.), (-1., -1.)];
@@ -633,24 +615,115 @@ impl App {
             .block(
                 Block::bordered()
                     .title(title.to_span().style(lb))
-                    // .title_bottom(Line::from("<- - + ->").left_aligned())
+                    .title_bottom(self.get_flashing_controls_text().left_aligned())
                     // current position and total duration
                     .title_bottom(
-                        Line::styled(format!("{:0>2}:{:0>5.2}", current_min, current_sec), tm)
+                        Line::styled(format!("{:0>2}:{:0>5.2}", current_min, current_sec), ct)
                             .centered(),
                     )
                     .title_bottom(
-                        Line::styled(format!("{:0>2}:{:0>5.2}", total_min, total_sec), tm)
+                        Line::styled(format!("{:0>2}:{:0>5.2}", total_min, total_sec), td)
                             .right_aligned(),
                     )
                     .title(upper_right_title)
                     .style(bd),
             )
             .style(wv)
-            .x_axis(Axis::default().bounds([0., 15. * 1000.]))
+            .x_axis(Axis::default().bounds([0., self.ui_settings.waveform_window * 1000.]))
             .y_axis(Axis::default().bounds([-1., 1.]));
 
         frame.render_widget(chart, area);
+    }
+
+    fn get_relative_playhead_pos(&self, samples_in_one_ms: u32) -> f64 {
+        // if at last 15 sec of the audion the playhead should move from the middle to the end of the chart
+        let total_samples = self.audio_file.mid_samples().len();
+        let chart_duration_seconds = self.ui_settings.waveform_window; // make a var not to hard code and be able to to add resizing waveform window if needed
+        let chart_middle_seconds = chart_duration_seconds / 2.0;
+
+        // calculate the absolute sample position where the playhead starts scrolling from the middle of the chart to the end
+        // this is when playback enters the last `chart_middle_seconds` (default is 7.5s) of the total audio duration.
+        let scroll_start_absolute_samples = total_samples
+            .saturating_sub((chart_middle_seconds * self.audio_file.sample_rate() as f64) as usize);
+
+        // calculate playhead's position relative to the start of this scroll phase
+        // since `self.waveform.playhead` is the absolute current playback position.
+        let relative_samples_in_scroll_phase = self
+            .waveform
+            .playhead
+            .saturating_sub(scroll_start_absolute_samples);
+
+        // map this relative sample position to the chart's X-axis range for the playhead.
+        // the conversion from samples to chart units (milliseconds) uses the same 1/samles_in_one_ms scale
+        // as other playhead positions in this function.
+        let mut chart_x_position = (chart_middle_seconds * 1000.)
+            + (relative_samples_in_scroll_phase as f64 / samples_in_one_ms as f64);
+
+        // Ensure the playhead does not exceed the chart's upper bound.
+        chart_x_position = f64::min(chart_x_position, chart_duration_seconds * 1000.);
+        chart_x_position
+    }
+
+    fn get_middle_playhead_pos(&self, samples_in_one_ms: u32) -> [(f64, f64); 2] {
+        [
+            (
+                f64::min(
+                    self.waveform.playhead as f64 / samples_in_one_ms as f64,
+                    1000. * self.ui_settings.waveform_window / 2.,
+                ),
+                1.,
+            ),
+            (
+                f64::min(
+                    self.waveform.playhead as f64 / samples_in_one_ms as f64,
+                    1000. * self.ui_settings.waveform_window / 2.,
+                ) + 0.01,
+                -1.,
+            ),
+        ]
+    }
+
+    fn get_flashing_controls_text(&self) -> Line<'_> {
+        let t = 100;
+        let s = Style::default()
+            .bg(self.ui_settings.theme.waveform.background.unwrap())
+            .fg(self.ui_settings.theme.waveform.controls.unwrap());
+        let hl = s.fg(self.ui_settings.theme.waveform.controls_highlight.unwrap());
+        let left_arrow = match self.ui_settings.left_arrow_timer {
+            Some(timer) if timer.elapsed().as_millis() < t => "<-".to_span().style(hl),
+            _ => "<-".to_span().style(s),
+        };
+        let right_arrow = match self.ui_settings.right_arrow_timer {
+            Some(timer) if timer.elapsed().as_millis() < t => "->".to_span().style(hl),
+            _ => "->".to_span().style(s),
+        };
+        let minus = match self.ui_settings.minus_sign_timer {
+            Some(timer) if timer.elapsed().as_millis() < t => "-".to_span().style(hl),
+            _ => "-".to_span().style(s),
+        };
+        let plus = match self.ui_settings.plus_sign_timer {
+            Some(timer) if timer.elapsed().as_millis() < t => "+".to_span().style(hl),
+            _ => "+".to_span().style(s),
+        };
+        // Line::from(format!(
+        //     "{} {} {:0>2}s {} {}",
+        //     left_arrow, minus, self.ui_settings.waveform_window, plus, right_arrow
+        // ))
+        Line::from(vec![
+            left_arrow,
+            " ".to_span(),
+            minus,
+            " ".to_span(),
+            format!(
+                "{:0>2}s",
+                self.ui_settings.waveform_window.to_span().style(s)
+            )
+            .into(),
+            " ".to_span(),
+            plus,
+            " ".to_span(),
+            right_arrow,
+        ])
     }
 
     fn render_fft_chart(&mut self, frame: &mut Frame, area: Rect) {
@@ -875,7 +948,9 @@ impl App {
             .enumerate()
             .map(|(i, (name, _dev))| ListItem::from(format!("[{}] {}", i + 1, name)))
             .collect();
-        let list = List::new(list_items).block(Block::bordered().title("Devices").style(s));
+        let list = List::new(list_items)
+            .style(s)
+            .block(Block::bordered().title("Devices").style(s));
 
         f.render_widget(list, area);
     }
@@ -933,7 +1008,7 @@ impl App {
             }
 
             // use ringbuf to analyze data if the `Mode` is not `Mode::Player`
-            if self.settings.mode == Mode::Microphone {
+            if matches!(self.settings.mode, Mode::Microphone) {
                 self.analyze_microphone_input();
             }
 
@@ -1015,21 +1090,24 @@ impl App {
         }
 
         //get waveform
+        let window = self.ui_settings.waveform_window as usize;
+        let half_window = self.ui_settings.waveform_window / 2.;
         let mid_samples_len = self.audio_file.mid_samples().len();
         self.waveform.playhead = pos;
         // if at zero load first 15 seconds to show
         if self.waveform.at_zero {
-            let waveform_samples = &self.audio_file.mid_samples()[0..15 * sr];
+            let waveform_samples = &self.audio_file.mid_samples()[0..window * sr];
             self.waveform.chart = Analyzer::get_waveform(waveform_samples, sr);
         }
-        let waveform_left_bound = pos.saturating_sub((7.5 * sr as f64) as usize);
-        let waveform_right_bound = usize::min(pos + (7.5 * sr as f64) as usize, mid_samples_len);
+        let waveform_left_bound = pos.saturating_sub((half_window * sr as f64) as usize);
+        let waveform_right_bound =
+            usize::min(pos + (half_window * sr as f64) as usize, mid_samples_len);
 
-        // if at end load last 15 seconds and dont scroll
+        // if at end load last `window` seconds and dont scroll
         if waveform_right_bound == mid_samples_len {
             self.waveform.at_end = true;
             let waveform_samples =
-                &self.audio_file.mid_samples()[mid_samples_len - 15 * sr..mid_samples_len];
+                &self.audio_file.mid_samples()[mid_samples_len - window * sr..mid_samples_len];
             self.waveform.chart = Analyzer::get_waveform(waveform_samples, sr);
         // if not at the beginning load 15 seconds and scroll
         } else if waveform_left_bound != 0 {
@@ -1095,6 +1173,7 @@ impl App {
                 if self.settings.mode == Mode::Player
                     && !(self.ui_settings.show_devices_list || self.ui_settings.show_explorer) =>
             {
+                self.ui_settings.right_arrow_timer = Some(Instant::now());
                 self.lufs = [-50.; 300];
                 self.analyzer.reset();
                 if let Err(_err) = self.player_command_tx.send(PlayerCommand::MoveRight) {
@@ -1105,6 +1184,7 @@ impl App {
                 if self.settings.mode == Mode::Player
                     && !(self.ui_settings.show_devices_list || self.ui_settings.show_explorer) =>
             {
+                self.ui_settings.left_arrow_timer = Some(Instant::now());
                 self.lufs = [-50.; 300];
                 self.analyzer.reset();
                 if let Err(_err) = self.player_command_tx.send(PlayerCommand::MoveLeft) {
@@ -1134,10 +1214,22 @@ impl App {
             // change mode. this will be replaced by normal settings selection tab
             // TODO normal settings popup window with a list of options
             KeyCode::Char('c') => {
-                self.settings.mode = if self.settings.mode == Mode::Microphone {
+                self.settings.mode = if matches!(self.settings.mode, Mode::Microphone) {
                     self.reset_charts();
+                    match self.audio_capture_stream.as_ref() {
+                        Some(stream) => {
+                            let _ = stream.pause();
+                        }
+                        None => (),
+                    }
                     Mode::Player
                 } else {
+                    match self.audio_capture_stream.as_ref() {
+                        Some(stream) => {
+                            let _ = stream.play();
+                        }
+                        None => (),
+                    }
                     Mode::Microphone
                 };
             }
@@ -1155,6 +1247,16 @@ impl App {
                     .set_cwd(config_dir().unwrap().join("soundscope"))
                     .unwrap();
                 self.ui_settings.show_themes_list = !self.ui_settings.show_themes_list;
+            }
+            KeyCode::Char('=') | KeyCode::Char('+') => {
+                self.ui_settings.plus_sign_timer = Some(Instant::now());
+                self.ui_settings.waveform_window =
+                    f64::max(self.ui_settings.waveform_window - 1., 1.);
+            }
+            KeyCode::Char('-') | KeyCode::Char('_') => {
+                self.ui_settings.minus_sign_timer = Some(Instant::now());
+                self.ui_settings.waveform_window =
+                    f64::min(self.ui_settings.waveform_window + 1., 15.);
             }
             _ => (),
         }
@@ -1637,7 +1739,7 @@ mod tests {
 
         theme.waveform.playhead = None;
         theme.waveform.highlight = None;
-        theme.waveform.time = None;
+        theme.waveform.current_time = None;
 
         theme.lufs.numbers = None;
 
@@ -1653,7 +1755,7 @@ mod tests {
 
         assert!(theme.waveform.playhead == Some(Color::LightRed));
         assert!(theme.waveform.highlight == Some(Color::LightRed));
-        assert!(theme.waveform.time == Some(Color::LightCyan));
+        assert!(theme.waveform.current_time == Some(Color::LightCyan));
 
         assert!(theme.lufs.numbers == Some(Color::LightCyan));